AOC2024Day14
My solutions to Advent of Code Day 14 😃

Part 2 I used the assumption that based on part 1 we would be looking for the smallest safety score in the pattern. If I had the time, I would maybe do it a different way that doesn't use that assumption like checking grids with a certain cluster score (many robots concentrated in a smallish area).